pyproject.toml
[project]
name = "ai-course"
version = "0.1.0"
description = ""
authors = [{name = "Sarah Kainec", email = "Sarah.Kainec@focusedlabs.io"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.8",
    "langchain-openai>=0.3.5",
    "python-dotenv>=1.0.1",
    "langchain-community>=0.3.17",
    "langsmith>=0.3.8",
    "deepgram-sdk>=3.10.0",
    "yt-dlp>=2025.1.26",
    "langchain-core>=0.3.74",
    "langchain-text-splitters>=0.3.9",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "langgraph-cli[inmem]>=0.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

